feat(es-dev-server): allow calling command-line-args separately
